`DefaultDecodable`: moved to `Misc/Codable`
Non functional change.

I was looking for these types and couldn't find them. Turns out it was under "FoundationExtensions", even though they're new definitions and not a foundation extension.
This creates a new `DefaultDecodable.swift`, and combines `JSONDecoder+Extensions` and `Decoder+Extensions`.
